# swagger/swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Phonebook API",
    "description": "API для работы с телефонным справочником",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/TS": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "number": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/TS/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "number": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "PhoneEntry": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Клочко Максим"
        },
        "number": {
          "type": "string",
          "example": "123-456-7890"
        }
      }
    },
    "PhoneEntries": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/PhoneEntry"
      }
    },
    "NewPhoneEntry": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Клочко Максим"
        },
        "number": {
          "type": "string",
          "example": "123-456-7890"
        }
      }
    },
    "UpdatedPhoneEntry": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Клочко Максим"
        },
        "number": {
          "type": "string",
          "example": "123-456-7890"
        }
      }
    }
  }
}